Modularize upper right portion of overview
