+++ client/src/components/Liam/upper-right/AddToCart.jsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from 'react';
+import ReactDOM from 'react-dom';
+// import 'bootstrap/dist/css/bootstrap.min.css'; don't need?
+
+
+const AddToCart = function () {
+
+  return (
+    <div className="add-to-cart">
+      Add to cart!
+    </div>
+  );
+};
+
+export default AddToCart;
