+++ client/src/components/Liam/upper-right/ProductInfo.jsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from 'react';
+import ReactDOM from 'react-dom';
+// import 'bootstrap/dist/css/bootstrap.min.css'; don't need?
+
+
+const ProductInfo = function () {
+
+  return (
+    <div className="product-info">
+      Product-info
+    </div>
+  );
+};
+
+export default ProductInfo;
