+++ client/src/components/Liam/upper-right/StyleSelector.jsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from 'react';
+import ReactDOM from 'react-dom';
+// import 'bootstrap/dist/css/bootstrap.min.css'; don't need?
+
+
+const StyleSelector = function () {
+
+  return (
+    <div className="style-selector">
+      Style-Selector
+    </div>
+  );
+};
+
+export default StyleSelector;
