+++ client/src/components/Liam/upper-right/UpperRight.jsx
@@ -1,21 +1,19 @@
 import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 // import 'bootstrap/dist/css/bootstrap.min.css'; don't need?
-
+import ProductInfo from './ProductInfo.jsx';
+import StyleSelector from './StyleSelector.jsx';
+import AddToCart from './AddToCart.jsx';
 
 const UpperRight = function () {
 
   return (
     <div className="upper-right">
-      <div className="product-info">
-        Product-info
-      </div>
-      <div className="style-selector">
-        Style Selector
-      </div>
-      <div className="add-to-cart">
-        Add to cart!
-      </div>
+      <ProductInfo />
+
+      <StyleSelector />
+
+      <AddToCart />
     </div>
   );
 };
